Move contents of preassembler.py into __init__
